ai-coding/pyproject.toml
[project]
name = "ai-coding"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.4",
    "pillow>=12.0.0",
]
